# BUG/PERF: Use EA in Index.get_value