--- a/pandas/tests/arrays/categorical/test_indexing.py
+++ b/pandas/tests/arrays/categorical/test_indexing.py
@@ -147,13 +147,28 @@ def test_mask_with_boolean_raises(index):
         s[idx]
 
 
-class NonCoercaibleCategorical(Categorical):
-    def __array__(self, dtype=None):
+@pytest.fixture
+def non_coercible_categorical(monkeypatch):
+    """
+    Monkeypatch Categorical.__array__ to ensure no implicit conversion.
+
+    Raises
+    ------
+    ValueError
+        When Categorical.__array__ is called.
+    """
+    # TODO(Categorical): identify other places where this may be
+    # useful and move to a conftest.py
+    def array(self, dtype=None):
         raise ValueError("I cannot be converted.")
 
+    with monkeypatch.context() as m:
+        m.setattr(Categorical, "__array__", array)
+        yield
 
-def test_series_at():
-    arr = NonCoercaibleCategorical(['a', 'b', 'c'])
+
+def test_series_at(non_coercible_categorical):
+    arr = Categorical(['a', 'b', 'c'])
     ser = Series(arr)
     result = ser.at[0]
     assert result == 'a'