@pep8speaks 2018-12-10
Hello @TomAugspurger! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`pandas/core/indexes/base.py`](https://github.com/TomAugspurger/pandas/blob/ac63e765bf5ed1efb4a315ae4ec8bf7f89c9600a/pandas/core/indexes/base.py) !

 - There are no PEP8 issues in the file [`pandas/tests/arrays/categorical/test_indexing.py`](https://github.com/TomAugspurger/pandas/blob/ac63e765bf5ed1efb4a315ae4ec8bf7f89c9600a/pandas/tests/arrays/categorical/test_indexing.py) !