@TomAugspurger 2018-12-10
All green.