@TomAugspurger 2018-12-10
Opened https://github.com/pandas-dev/pandas/issues/24207 for followup.